SP-2570 Optimise prepare_facet memory usage
This relatively simple function needs to handle the largest blocks of
data (facets before and after preparation), and is therefore likely
responsible for the "spikes" of memory usage we have been seeing.

This modifies the function such that when the size of the input goes
beyond ~1MB, it starts pre-allocating the entire output, then filling
it incrementally. This should theoretically minimise the amount of
memory we utilise.