SP-2570 Implement sending unshifted contributions
The idea here is that if we construct the contributions to *only* be
aliased but *not* shifted relative to the zero point in image and
frequency, we can get rid of the restrictions on subgrid and facet
offsets - they actually don't interact with each other any more.

This only works with the new-style "flat m" approach, so we are also
starting to remove support for old-style configurations. Net result is
that we can remove any mention of

* Nx (always N//yN_size now due to unshifted contributions)
* Ny (always N//xM_size now due to unshifted contributions)
* yP (always yN_size now due to new-style flat-m configs)
* and m (always trivial top-hat now, as "flat-m" suggests)