SP-2570 Refactor core interface
This removes the "mask" parameter from the core interfaces - it can be easily
applied externally.

More significantly, this adds an "out" parameter to all processing functions so
that we can - at least in principle - reduce the amount of allocation
done. This is chiefly useful for ska-sdp-func core, as it eliminates allocation
entirely. However, for the purpose of consistency, this also adds the
equivalent feature to the numpy-based core processing functions.